separate rate history chart as its own component
